add bodyClassname and reorganize modal
