+++ packages/modal/package.json
@@ -28,6 +28,7 @@
     "@leafygreen-ui/palette": "^4.0.3",
     "@leafygreen-ui/portal": "^4.1.1",
     "@leafygreen-ui/tokens": "^2.0.2",
+    "@leafygreen-ui/typography": "^16.3.0",
     "focus-trap-react": "^8.10.0",
     "polished": "^4.2.2",
     "prop-types": "^15.8.1",
@@ -36,7 +37,7 @@
   "devDependencies": {
     "@leafygreen-ui/select": "^10.3.1",
     "@leafygreen-ui/button": "^20.0.4",
-    "@leafygreen-ui/typography": "^16.2.1",
+    "@leafygreen-ui/typography": "^16.3.0",
     "@leafygreen-ui/copyable": "^8.0.7",
     "@leafygreen-ui/code": "^14.2.2"
   },
