+++ packages/modal/src/Footer/Footer.styles.ts
