+++ packages/modal/src/Footer/Footer.types.ts
