+++ packages/modal/src/Footer/index.ts
@@ -1,0 +1,1 @@
+export { default as Footer } from './Footer';
