+++ packages/modal/src/Modal/index.ts
@@ -1,0 +1,3 @@
+export { default } from './Modal';
+export { CloseIconColor, type ModalProps, ModalSize } from './Modal.types';
+export { default as ModalView } from './ModalView';
