+++ packages/modal/src/index.ts
@@ -1,5 +1,8 @@
-export { default as Footer } from './Footer/Footer';
-export { default } from './Modal/Modal';
-export { default as ModalView } from './Modal/ModalView';
-export type { ModalProps } from './Modal/types';
-export { CloseIconColor, ModalSize } from './Modal/types';
+export { Footer } from './Footer';
+export {
+  CloseIconColor,
+  default,
+  type ModalProps,
+  ModalSize,
+  ModalView,
+} from './Modal';
