+++ packages/modal/tsconfig.json
@@ -36,6 +36,9 @@
     },
     {
       "path": "../leafygreen-provider"
+    },
+    {
+      "path": "../typography"
     }
   ]
 }
