+++ packages/typography/src/Body/Body.tsx
@@ -10,6 +10,7 @@ import {
 
 import {
   baseTypographyStyles,
+  bodyClassName,
   bodyTypeScaleStyles,
   defaultTextColor,
 } from '../styles';
@@ -60,6 +61,7 @@ const Body = Polymorphic<BodyProps>(
           bodyTypeScaleStyles[baseFontSize],
           defaultTextColor[theme],
           fontWeight,
+          bodyClassName,
           className,
         )}
         {...rest}
